refactor: create separate binary crates and move TrinConfig to trin crate
