+++ bin/historical-batch/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "historical-batch"
+authors.workspace = true
+categories.workspace = true
+edition.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+dirs = "5.0.1"
+e2store.workspace = true
+ethereum_ssz.workspace = true
+ethportal-api.workspace = true
+regex = "1.10.2"
+reqwest.workspace = true
+tokio.workspace = true
+tree_hash.workspace = true
