+++ bin/poll-latest/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "poll-latest"
+authors.workspace = true
+categories.workspace = true
+edition.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+alloy = { workspace = true, features = ["provider-ipc", "provider-ws", "pubsub", "reqwest", "rpc-types"] }
+anyhow.workspace = true
+clap.workspace = true
+ethportal-api.workspace = true
+futures.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+trin-utils.workspace = true
+url.workspace = true
