+++ bin/test-providers/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "test-providers"
+authors.workspace = true
+categories.workspace = true
+edition.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+clap.workspace = true
+ethereum_ssz.workspace = true
+ethportal-api.workspace = true
+lazy_static.workspace = true
+portal-bridge.workspace = true
+rand.workspace = true
+reqwest.workspace = true
+serde_json = { workspace = true, features = ["preserve_order"]}
+tokio.workspace = true
+tracing.workspace = true
+trin-utils.workspace = true
+trin-validation.workspace = true
+url.workspace = true
