+++ ethportal-api/src/lib.rs
@@ -4,11 +4,10 @@
 #![warn(clippy::unwrap_used)]
 #![warn(clippy::uninlined_format_args)]
 
-#[macro_use]
 extern crate lazy_static;
 
 mod beacon;
-mod dashboard;
+pub mod dashboard;
 pub mod discv5;
 mod eth;
 mod history;
