+++ portalnet/src/config.rs
@@ -1,17 +1,14 @@
 use std::net::SocketAddr;
 
 use alloy::primitives::B256;
-use ethportal_api::types::{
-    bootnodes::Bootnodes,
-    cli::{TrinConfig, DEFAULT_UTP_TRANSFER_LIMIT},
-    enr::Enr,
-    network::Network,
-};
+use ethportal_api::types::{enr::Enr, network::Network};
+
+use crate::{bootnodes::Bootnodes, constants::DEFAULT_UTP_TRANSFER_LIMIT};
 
 /// Capacity of the cache for observed `NodeAddress` values.
 /// Provides capacity for 32 full k-buckets. This capacity will be shared among all active portal
 /// subnetworks.
-const NODE_ADDR_CACHE_CAPACITY: usize = discv5::kbucket::MAX_NODES_PER_BUCKET * 32;
+pub const NODE_ADDR_CACHE_CAPACITY: usize = discv5::kbucket::MAX_NODES_PER_BUCKET * 32;
 
 #[derive(Clone)]
 pub struct PortalnetConfig {
@@ -42,23 +39,6 @@ impl Default for PortalnetConfig {
             disable_poke: false,
             trusted_block_root: None,
             utp_transfer_limit: DEFAULT_UTP_TRANSFER_LIMIT,
-        }
-    }
-}
-
-impl PortalnetConfig {
-    pub fn new(trin_config: &TrinConfig, private_key: B256) -> Self {
-        Self {
-            external_addr: trin_config.external_addr,
-            private_key,
-            listen_port: trin_config.discovery_port,
-            bootnodes: trin_config.bootnodes.to_enrs(trin_config.network.network()),
-            no_stun: trin_config.no_stun,
-            no_upnp: trin_config.no_upnp,
-            node_addr_cache_capacity: NODE_ADDR_CACHE_CAPACITY,
-            disable_poke: trin_config.disable_poke,
-            trusted_block_root: trin_config.trusted_block_root,
-            utp_transfer_limit: trin_config.utp_transfer_limit,
         }
     }
 }
